+++ src/main/java/com/houseofpizza/repository/specification/PizzaToOrderSpecification.java
@@ -9,12 +9,12 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 public class PizzaToOrderSpecification {
 
-    public Specification<PizzaToOrder> withIdOrderEqualTo(final Integer orderNumber) {
+    public Specification<PizzaToOrder> withIdOrderEqualTo(final Long orderNumber) {
         return ((root, criteriaQuery, criteriaBuilder) -> criteriaBuilder
             .equal(root.get(PizzaToOrder.Fields.idOrder), orderNumber));
     }
 
-    public Specification<PizzaToOrder> withIdStatusEqualTo(final Integer idStatus) {
+    public Specification<PizzaToOrder> withIdStatusEqualTo(final Long idStatus) {
         return ((root, criteriaQuery, criteriaBuilder) -> criteriaBuilder
             .equal(root.get(PizzaToOrder.Fields.idStatus), idStatus));
     }
